Add cross-platform filesystem adapter coverage
