+++ codex-rs/exec-server/tests/file_system_windows.rs
@@ -1,0 +1,51 @@
+#![cfg(windows)]
+
+mod common;
+
+#[path = "file_system/shared.rs"]
+mod shared;
+#[path = "file_system/support.rs"]
+mod support;
+
+use std::path::Path;
+use std::process::Command;
+
+use anyhow::Result;
+use test_case::test_case;
+
+fn create_directory_junction(target: &Path, alias: &Path) -> Result<()> {
+    let output = Command::new("cmd")
+        .args(["/C", "mklink", "/J"])
+        .arg(alias)
+        .arg(target)
+        .output()?;
+    if !output.status.success() {
+        anyhow::bail!(
+            "mklink /J failed: stdout={} stderr={}",
+            String::from_utf8_lossy(&output.stdout).trim(),
+            String::from_utf8_lossy(&output.stderr).trim()
+        );
+    }
+    Ok(())
+}
+
+#[test_case(false ; "local")]
+#[test_case(true ; "remote")]
+#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+async fn file_system_canonicalize_resolves_directory_junction(use_remote: bool) -> Result<()> {
+    shared::assert_canonicalize_resolves_directory_alias(use_remote, create_directory_junction)
+        .await
+}
+
+#[test_case(false ; "local")]
+#[test_case(true ; "remote")]
+#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+async fn file_system_sandboxed_canonicalize_resolves_directory_junction(
+    use_remote: bool,
+) -> Result<()> {
+    shared::assert_sandboxed_canonicalize_resolves_directory_alias(
+        use_remote,
+        create_directory_junction,
+    )
+    .await
+}
